fix regression -Werror messages
